- RqidLink click handler now only in RqidLink
