+++ src/actors/rqgActorSheet.ts
@@ -1048,21 +1048,7 @@ export class RqgActorSheet extends ActorSheet<
     });
 
     // Handle rqid links
-    $(this.form!)
-      .find("[data-rqid-link]")
-      .each((i: number, el: HTMLElement) => {
-        const rqid = getRequiredDomDataset($(el), "rqid");
-        el.addEventListener("click", async () => {
-          const rqidItem = await Rqid.fromRqid(rqid);
-          if (rqidItem) {
-            rqidItem.sheet?.render(true);
-          } else {
-            ui.notifications?.warn(
-              localize("RQG.Item.Notification.RqidFromLinkNotFound", { rqid: rqid }) // TODO More generic notification
-            );
-          }
-        });
-      });
+    RqidLink.addRqidLinkClickHandlers($(this.form!));
   }
 
   static confirmItemDelete(actor: RqgActor, itemId: string): void {
