+++ src/items/RqgItemSheet.ts
@@ -139,21 +139,7 @@ export class RqgItemSheet<
       });
 
     // Handle rqid links
-    $(this.form!)
-      .find("[data-rqid-link]")
-      .each((i: number, el: HTMLElement) => {
-        const rqid = getRequiredDomDataset($(el), "rqid");
-        el.addEventListener("click", async () => {
-          const rqidItem = await Rqid.fromRqid(rqid);
-          if (rqidItem) {
-            rqidItem.sheet?.render(true);
-          } else {
-            ui.notifications?.warn(
-              localize("RQG.Item.Notification.RqidFromLinkNotFound", { rqid: rqid })
-            );
-          }
-        });
-      });
+    RqidLink.addRqidLinkClickHandlers($(this.form!));
 
     $(this.form!)
       .find("[data-delete-from-property]")
